refactor: 체스게임 컨트롤러와 관련된 코드 이동 및 url 수정
