+++ src/main/java/chess/controller/LobbyController.java
@@ -1,12 +1,9 @@
 package chess.controller;
 
-import chess.dto.ChessGameRequest;
 import chess.service.ChessGameService;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ModelAttribute;
-import org.springframework.web.bind.annotation.PostMapping;
 
 
 @Controller
@@ -24,9 +21,4 @@ public class LobbyController {
         return "index";
     }
 
-    @PostMapping("/create-chess-game")
-    public String createChessGame(@ModelAttribute ChessGameRequest chessGameRequest) {
-        chessGameService.create(chessGameRequest);
-        return "redirect:/";
-    }
 }
