+++ src/main/resources/templates/index.hbs
@@ -9,7 +9,7 @@
 <div class="lobby">
     <h1>체스 게임 대기실</h1>
     <div class="create-room">
-        <form action="/create-chess-game" method="post">
+        <form action="/chess-game" method="post">
             <input type="hidden" name="id" value="{{id}}"/>
             <input type="text" name="name" placeholder="방 이름을 입력하세요."/>
             <input type="password" name="password" placeholder="비밀번호를 입력해주세요."/>
